Complete rework of Event class part 2
- Turned events.py package into events/ module
- Working state
- Before and after "snapshots" are stored as attributes of each agent in
  each event
- This makes backtracking and retrospection possible. For example, one
  could go to an event, resolve it a _different_ way, and then
  re-simulate
- Event objects no longer contain references to the active objects in a
  system. Instead, all of the snapshots are deepcopies. This is great,
  because the variables don't silently change as the system evolves
