+++ pooltool/system/datatypes.py
@@ -8,7 +8,14 @@ import numpy as np
 import pooltool.physics as physics
 import pooltool.utils as utils
 from pooltool.error import ConfigError
-from pooltool.events import Event, event_resolvers, filter_ball, stick_ball_collision
+from pooltool.events import (
+    AgentType,
+    Event,
+    EventType,
+    event_resolvers,
+    filter_ball,
+    stick_ball_collision,
+)
 from pooltool.objects.ball.datatypes import Ball, BallHistory, BallState
 from pooltool.objects.cue.datatypes import Cue
 from pooltool.objects.table.components import Pocket
@@ -144,15 +151,13 @@ class System:
 
                     # We need to get the ball's outgoing state from the event. We'll
                     # evolve the system from this state.
-                    if events[count].event_type.is_transition():
-                        state = events[count].agents[0].final.state
-                    elif events[count].event_type.is_collision():
-                        if isinstance(events[count].agents[0].final, Ball):
-                            state = events[count].agents[0].final.state
-                        else:
-                            state = events[count].agents[1].final.state
+                    for agent in events[count].agents:
+                        if agent.matches(ball):
+                            state = agent.get_final().state
+                            break
                     else:
-                        raise NotImplementedError(f"Can't handle {events[count]}")
+                        raise ValueError("No agents in event match ball")
+
                     rvw, s = state.rvw, state.s
 
                     # Since this event occurs between two timestamps, we won't be
@@ -179,7 +184,7 @@ class System:
             # Attach the newly created history to the ball
             ball.history_cts = history
 
-    def evolve(self, dt):
+    def evolve(self, dt: float):
         """Evolves current ball an amount of time dt
 
         FIXME This is very inefficent. each ball should store its natural trajectory
@@ -202,15 +207,31 @@ class System:
             ball.state.set(rvw, s=s, t=(self.t + dt))
 
     def resolve_event(self, event: Event) -> None:
-        event.resolve()
+        if event.event_type == EventType.NONE:
+            return
+
+        # The initial states need to be snapshotted
+        for agent in event.agents:
+            if agent.agent_type == AgentType.CUE:
+                agent.set_initial(self.cue)
+            elif agent.agent_type == AgentType.BALL:
+                agent.set_initial(self.balls[agent.id])
+            elif agent.agent_type == AgentType.POCKET:
+                agent.set_initial(self.table.pockets[agent.id])
+            elif agent.agent_type == AgentType.LINEAR_CUSHION_SEGMENT:
+                agent.set_initial(self.table.cushion_segments.linear[agent.id])
+            elif agent.agent_type == AgentType.CIRCULAR_CUSHION_SEGMENT:
+                agent.set_initial(self.table.cushion_segments.circular[agent.id])
+
+        event = event_resolvers[event.event_type](event)
 
         # The final states of the agents are solved, but the system objects still need
         # to be updated with these states.
         for agent in event.agents:
-            final = agent.final
+            final = agent.get_final()
             if isinstance(final, Ball):
                 self.balls[final.id].state = final.state
-            if isinstance(final, Pocket):
+            elif isinstance(final, Pocket):
                 self.table.pockets[final.id] = final
 
     def reset_balls(self):
